Simplify register_term: make recursive, use get_subterms uniformly
- Egraph::register_term is now recursive (registers subterms + predecessor
  edges internally). Solver only calls it once at the top level.
- SolverState::insert_predecessor simplified to solver bookkeeping
  (arithmetic tracking, quantifier registration) + one register_term call.
- Merged App/Ite function_maps logic into single get_subterms call.
- from_quantifier derived from decision_level > 0 internally.
- Removed parent and disequalities params from the public API.
